Add jwt middleware and create protected routed

+++ models/event.go
@@ -3,10 +3,10 @@ package models
 import "time"
 
 type Event struct {
-	Id          int
+	Id          int64
 	Name        string    `binding:"required"`
 	Description string    `binding:"required"`
 	Location    string    `binding:"required"`
 	DateTime    time.Time `binding:"required"`
-	UserId      int
+	UserId      int64
 }
